chore: rename types to type_defs
