+++ python/x402_a2a/executors/base.py
@@ -15,7 +15,7 @@
 
 from abc import ABC, abstractmethod
 
-from ..types import (
+from ..type_defs import (
     AgentExecutor,
     RequestContext,
     EventQueue,
